requirements.txt
apache-airflow==2.9.*
airflow-ai-sdk>=0.1.0
pydantic-ai>=0.0.13
httpx>=0.27
colorlog>=6.8
apache-airflow-providers-slack>=8.0